+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/AvaCorpAvilan.java
@@ -323,7 +323,7 @@ public final class AvaCorpAvilan extends Sistema {
         final PosicaoPneuMapper posicaoPneuMapper = new PosicaoPneuMapper(
                 dao.getPosicoesPneuAvilanProLogByCodTipoVeiculoAvilan(codTipoVeiculoAvilan));
 
-        final Afericao afericao = AvaCorpAvilanConverter.convert(posicaoPneuMapper, afericaoFiltro);
+        final Afericao afericao = AvaCorpAvilanConverter.convert(posicaoPneuMapper, afericaoFiltro, codUnidade);
 
         final Short codDiagrama = dao.getCodDiagramaVeiculoProLogByCodTipoVeiculoAvilan(codTipoVeiculoAvilan);
         final Optional<DiagramaVeiculo> optional = getIntegradorProLog().getDiagramaVeiculoByCodDiagrama(codDiagrama);
@@ -366,7 +366,7 @@ public final class AvaCorpAvilan extends Sistema {
                 getCpf(),
                 getDataNascimento());
 
-        return AvaCorpAvilanConverter.convertAfericoes(afericoes.getAfericaoFiltro());
+        return AvaCorpAvilanConverter.convertAfericoes(afericoes.getAfericaoFiltro(), codUnidade);
     }
 
     @NotNull